AI-agent-pocs/claud-poc/requirements.txt
langchain  
langchain-community 
langchain-anthropic
python-dotenv
pydantic